+++ killtracker/admin.py
@@ -1,18 +1,23 @@
 from django.contrib import admin
 from django.db.models import Q
 from django.db.models.functions import Lower
-from django.http import HttpResponseRedirect
-from django.shortcuts import render
+from django.http import HttpResponse, HttpResponseRedirect
+from django.shortcuts import get_object_or_404, render
 from django.utils.safestring import mark_safe
-from eveuniverse.models import EveGroup, EveType
+from eveuniverse.models import EveGroup
 
 from allianceauth.eveonline.models import EveAllianceInfo, EveCorporationInfo
 
 from . import tasks
-from .constants import EveCatagoryId, EveGroupId
+from .constants import (
+    SESSION_KEY_TOOGLE_NPC,
+    SESSION_KEY_USES_NPC,
+    EveCategoryId,
+    EveGroupId,
+)
 from .core.killmails import Killmail
 from .forms import TrackerAdminForm, TrackerAdminKillmailIdForm, field_nice_display
-from .models import Tracker, Webhook
+from .models import EveTypePlus, Tracker, Webhook
 
 
 @admin.register(Webhook)
@@ -207,6 +212,25 @@ class TrackerAdmin(admin.ModelAdmin):
             "ping_groups",
         )
 
+    def change_view(
+        self, request, object_id, form_url="", extra_context=None
+    ) -> HttpResponse:
+        extra_context = extra_context or {}
+        tracker = get_object_or_404(Tracker, pk=object_id)
+        is_using_npc = (
+            tracker.require_attackers_ship_types.filter(
+                eve_group__eve_category_id=EveCategoryId.ENTITY
+            ).exists()
+            or tracker.require_attackers_ship_groups.filter(
+                eve_category_id=EveCategoryId.ENTITY
+            ).exists()
+        )
+        request.session[SESSION_KEY_USES_NPC] = is_using_npc
+        extra_context["is_using_npc"] = is_using_npc
+        return super().change_view(
+            request, object_id, form_url, extra_context=extra_context
+        )
+
     def _color(self, obj):
         html = (
             f'<input type="color" value="{obj.color}" disabled>' if obj.color else "-"
@@ -326,7 +350,7 @@ class TrackerAdmin(admin.ModelAdmin):
         form = TrackerAdminKillmailIdForm(initial=initial)
         return render(
             request,
-            "admin/killtracker_killmail_id.html",
+            "admin/killtracker/tracker/killmail_test.html",
             {
                 "form": form,
                 "title": "Load Test Killmail for Tracker",
@@ -336,6 +360,9 @@ class TrackerAdmin(admin.ModelAdmin):
 
     def formfield_for_manytomany(self, db_field, request, **kwargs):
         """overriding this formfield to have sorted lists in the form"""
+        show_npc_types = request.session.get(
+            SESSION_KEY_USES_NPC, False
+        ) or request.session.get(SESSION_KEY_TOOGLE_NPC, False)
         if db_field.name in {
             "exclude_attacker_alliances",
             "require_attacker_alliances",
@@ -355,22 +382,34 @@ class TrackerAdmin(admin.ModelAdmin):
                 Lower("corporation_name")
             )
         elif db_field.name == "require_attackers_ship_groups":
-            kwargs["queryset"] = EveGroup.objects.filter(
+            qs = EveGroup.objects.filter(
                 eve_category_id__in=[
-                    EveCatagoryId.STRUCTURE,
-                    EveCatagoryId.SHIP,
-                    EveCatagoryId.FIGHTER,
+                    EveCategoryId.STRUCTURE,
+                    EveCategoryId.SHIP,
+                    EveCategoryId.FIGHTER,
                 ],
                 published=True,
-            ).order_by(Lower("name"))
+            )
+            if show_npc_types:
+                qs = (
+                    qs
+                    | EveGroup.objects.filter(
+                        eve_category_id=EveCategoryId.ENTITY
+                    ).filter(
+                        eve_types__mass__gt=0,
+                        eve_types__volume__gt=0,
+                        eve_types__capacity__gt=0,
+                    )
+                ).distinct()
+            kwargs["queryset"] = qs.order_by(Lower("name"))
         elif db_field.name == "require_victim_ship_groups":
             kwargs["queryset"] = EveGroup.objects.filter(
                 (
                     Q(
                         eve_category_id__in=[
-                            EveCatagoryId.STRUCTURE,
-                            EveCatagoryId.SHIP,
-                            EveCatagoryId.FIGHTER,
+                            EveCategoryId.STRUCTURE,
+                            EveCategoryId.SHIP,
+                            EveCategoryId.FIGHTER,
                         ]
                     )
                     & Q(published=True)
@@ -379,35 +418,35 @@ class TrackerAdmin(admin.ModelAdmin):
                 | Q(id=EveGroupId.ORBITAL_INFRASTRUCTURE)
             ).order_by(Lower("name"))
         elif db_field.name == "require_attackers_ship_types":
-            kwargs["queryset"] = (
-                EveType.objects.select_related("eve_group")
-                .filter(
-                    eve_group__eve_category_id__in=[
-                        EveCatagoryId.STRUCTURE,
-                        EveCatagoryId.SHIP,
-                        EveCatagoryId.FIGHTER,
-                    ],
-                    published=True,
-                )
-                .order_by(Lower("name"))
+            qs = EveTypePlus.objects.filter(
+                eve_group__eve_category_id__in=[
+                    EveCategoryId.STRUCTURE,
+                    EveCategoryId.SHIP,
+                    EveCategoryId.FIGHTER,
+                ],
+                published=True,
             )
+            if show_npc_types:
+                qs = qs | EveTypePlus.objects.filter(
+                    eve_group__eve_category_id=EveCategoryId.ENTITY,
+                    mass__gt=0,
+                    volume__gt=0,
+                    capacity__gt=0,
+                )
+            kwargs["queryset"] = qs.order_by(Lower("name"))
         elif db_field.name == "require_victim_ship_types":
-            kwargs["queryset"] = (
-                EveType.objects.select_related("eve_group")
-                .filter(
-                    (
-                        Q(
-                            eve_group__eve_category_id__in=[
-                                EveCatagoryId.STRUCTURE,
-                                EveCatagoryId.SHIP,
-                                EveCatagoryId.FIGHTER,
-                            ]
-                        )
-                        & Q(published=True)
+            kwargs["queryset"] = EveTypePlus.objects.filter(
+                (
+                    Q(
+                        eve_group__eve_category_id__in=[
+                            EveCategoryId.STRUCTURE,
+                            EveCategoryId.SHIP,
+                            EveCategoryId.FIGHTER,
+                        ]
                     )
-                    | (Q(eve_group_id=EveGroupId.MINING_DRONE) & Q(published=True))
-                    | Q(eve_group_id=EveGroupId.ORBITAL_INFRASTRUCTURE)
+                    & Q(published=True)
                 )
-                .order_by(Lower("name"))
-            )
+                | (Q(eve_group_id=EveGroupId.MINING_DRONE) & Q(published=True))
+                | Q(eve_group_id=EveGroupId.ORBITAL_INFRASTRUCTURE)
+            ).order_by(Lower("name"))
         return super().formfield_for_manytomany(db_field, request, **kwargs)