+++ killtracker/tests/testdata/create_eveuniverse.py
@@ -7,10 +7,22 @@ from . import test_data_filename
 class CreateEveUniverseTestData(TestCase):
     def test_create_testdata(self):
         testdata_spec = [
-            ModelSpec("EveFaction", ids=[500001]),
+            ModelSpec("EveFaction", ids=[500001, 500004]),
             ModelSpec(
                 "EveType",
-                ids=[603, 621, 638, 2488, 2977, 3756, 11379, 16238, 34562, 37483],
+                ids=[
+                    603,
+                    621,
+                    638,
+                    2488,
+                    2977,
+                    3756,
+                    11379,
+                    16238,
+                    34562,
+                    37483,
+                    23320,
+                ],
             ),
             ModelSpec(
                 "EveSolarSystem", ids=[30001161, 30004976, 30004984, 30045349, 31000005]